arm disarm test changes

+++ Sabin-Code/src/protocol.h
@@ -40,7 +40,7 @@ private:
 
 	void msp_request(uint8_t);
 
-	void msp_command(uint8_t, uint8_t, int*);
+	void msp_command(uint8_t, uint8_t, uint16_t*);
 
 	std::string read();
 
@@ -53,9 +53,9 @@ public:
 
 	std::string request_data(uint8_t);
 
-	std::string request_data(uint8_t, uint8_t, int*);
+	std::string request_data(uint8_t, uint8_t, uint16_t*);
 
-	void set_data(uint8_t, uint8_t, int*);
+	void set_data(uint8_t, uint8_t, uint16_t*);
 	 
 };
